# Tude-Web-App-master/backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Étude+ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/checkout:
    post:
      operationId: checkout
      tags: [transactions]
      summary: Checkout payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
components:
  schemas:
...
    CheckoutRequest:
      type: object
      properties:
        classId:
          type: number
        sessionId:
          type: number
          nullable: true
      required: [classId]
...
    CheckoutResponse:
      type: object
      properties:
        transactionId:
          type: number
        amount:
          type: number
        platformFee:
          type: number
        professorAmount:
          type: number
        classTitle:
          type: string
        professorName:
          type: string
      required: [transactionId, amount, platformFee, professorAmount, classTitle, professorName]
...